fix: Edit Ctrl+Backspace clearing undo buffer
Calling SetText clears the undo buffer. So instead switched to
REPLACESEL with wParam set to FALSE which makes it not clear the
undo buffer. I'm also only selecting the bits to be deleted rather
than the whole text as that way when you undo it selects only those
bits, which is more in line with usual undo behaviour.